vm: fix some size_t/uint64_t issues (14)

+++ src/cshared/asm_stream.h
@@ -9,7 +9,7 @@ struct BaseReader
     {
     }
 
-    size_t position() const
+    uint64_t position() const
     {
         return _position;
     }
